+++ source/extensions/lightspeed.event.layers_cleanup/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.0.0] - 2023-10-23
+### Added
+- Remove invalid sublayers on stage loaded or sublayer changed
+
+### Removed
+- Don't cleanup metadata on stage loaded
+
 ## [1.1.0] - 2023-10-10
 ### Changed
 - Code cleanup